# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PolariixApi",
        "description": "Documentation complète de l'API - Tous les endpoints sont accessibles ici",
        "contact": {
            "name": "Support API",
            "email": "contact@votreapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur API"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "description": "Crée un nouveau compte utilisateur. Un email de vérification est envoyé automatiquement. Le compte reste inactif jusqu'à la vérification de l'email.",
                "operationId": "6796c44e1f6079c411abbfed066c5907",
                "requestBody": {
                    "description": "Données d'inscription de l'utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom complet de l'utilisateur",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Adresse email unique",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe (minimum 8 caractères)",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "Password123!"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmation du mot de passe",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "phone": {
                                        "description": "Numéro de téléphone (optionnel)",
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+229 97 00 00 00",
                                        "nullable": true
                                    },
                                    "role": {
                                        "description": "Rôle de l'utilisateur (par défaut: user)",
                                        "type": "string",
                                        "enum": [
                                            "user",
                                            "admin",
                                            "superadmin"
                                        ],
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie. Email de vérification envoyé.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGc..."
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inscription réussie. Un email de vérification a été envoyé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Vérification de l'email utilisateur",
                "description": "Valide l'adresse email de l'utilisateur via le lien reçu par email. Active automatiquement le compte après vérification.",
                "operationId": "19aaa873091d0a21743efdf13d858d14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash de vérification de l'email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "a1b2c3d4e5f6..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email vérifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email vérifié avec succès. Statut activé."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lien de vérification invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lien de vérification invalide."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/enable-2fa": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Activer l'authentification à deux facteurs (2FA)",
                "description": "Active la 2FA pour l'utilisateur connecté. Le compte passe en statut 'inactive' et un email de validation est envoyé.",
                "operationId": "d3104bdb8c67feeb6175ef933dab217f",
                "responses": {
                    "200": {
                        "description": "2FA activée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "2FA activée, un email de validation vous a été envoyé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Réponse d'erreur standard",
                "description": "Structure standard pour les réponses d'erreur",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue"
                    },
                    "code": {
                        "type": "integer",
                        "example": 400
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Erreur de validation",
                "description": "Structure pour les erreurs de validation",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                        "properties": {
                            "errors": {
                                "type": "object",
                                "example": {
                                    "email": [
                                        "L'email est invalide"
                                    ]
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via JWT Token. Entrez votre token sans le préfixe 'Bearer'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}